Use get_api_class for IsApiRequest
